rename to nonce_extraction
