+++ runtime/src/lib.rs
@@ -16,6 +16,7 @@ pub mod genesis_utils;
 pub mod installed_scheduler_pool;
 pub mod loader_utils;
 pub mod non_circulating_supply;
+pub mod nonce_extraction;
 pub mod prioritization_fee;
 pub mod prioritization_fee_cache;
 pub mod rent_collector;
@@ -34,7 +35,6 @@ pub mod stake_weighted_timestamp;
 pub mod stakes;
 pub mod static_ids;
 pub mod status_cache;
-pub mod svm_message_ext;
 pub mod transaction_batch;
 pub mod vote_sender_types;
